refactor: 피드 페이지 코로케이션 적용
