+++ src/pages/feed/components/BestIdeaCardListSection/BestIdeaCardListSection.tsx
@@ -2,7 +2,7 @@ import styled from '@emotion/styled';
 import { Spacer, Text } from 'concept-be-design-system';
 
 import BestIdeaCard from '../../../../components/Card/BestIdeaCard/BestIdeaCard';
-import { useBestIdeasQuery } from '../../../../hooks/queries/useBestIdeasQuery';
+import { useBestIdeasQuery } from '../../hooks/queries/useBestIdeasQuery';
 
 const BestIdeaCardListSection = () => {
   const { bestIdeas } = useBestIdeasQuery();
